use lists of databases, usernames and passwords
This deprecates the variables MONGODB_DATABASE, MONGODB_USERNAME and
MONGODB_PASSWORD in favor for environment variables with a trailing S.
The new variables can be used to initialize several databases, with as
many users as necessary the first time a container is run. In these
variables, comma and semi-colon signs can be used as a delimiter. The
use of these characters is compatible with other Bitnami images. The
implementation still recognizes the old variables, but will print a
deprecation warning.