revert to ALLOW_EMPTY_PASSWORD and allow empty db
This reverts back to using the variable `ALLOW_EMPTY_PASSWORD` (without
the trailing `S`).

It is also possible to run with an empty variable `MONGODB_DATABASES`,
in which case the users and their respective passwords specified
through `MONGODB_USERNAMES` and `MONGODB_PASSWORDS` will be created in the
default database.

Finally, the implementation has been simplified through keeping a
reference to the old variables in file mongodb-envs.sh.