fix proper validation and add warning
This fixes the validation so that as soon as authorization is turned on,
the root password specified in `MONGODB_ROOT_PASSWORD` cannot be empty
unless `ALLOW_EMPTY_PASSWORD` was set.

The validation also shows a warning for users that will not be created
whenever their passwords are empty. MongoDB refuses to create users with
empty passwords. The validation prints a warning, not an error to keep
compatibility with prior versions.